OneDrive/Desktop/CyberEye-3.0/backend/requirements.txt
flask
flask-cors
requests
python-dotenv
sqlite3